+++ GymMaster/src/java/gym/TiposSangre.java
@@ -13,18 +13,18 @@ public interface TiposSangre {
     /**
      * Constante que representa el tipo de sangre A
      */
-    public static final byte TIPO_A = 0;
+    public static final short TIPO_A = 0;
     /**
      * Constante que representa el tipo de sangre B
      */
-    public static final byte TIPO_B = 1;
+    public static final short TIPO_B = 1;
     /**
      * Constante que representa el tipo de sangre AB
      */
-    public static final byte TIPO_AB = 2;
+    public static final short TIPO_AB = 2;
     /**
      * Constante que representa el tipo de sangre O
      */
-    public static final byte TIPO_O = 3;
+    public static final short TIPO_O = 3;
     
 }
